--- modin/experimental/engines/gandiva_on_ray/io.py
@@ -1,11 +1,0 @@
-from modin.data_management.query_compiler import GandivaQueryCompiler
-from modin.engines.base.io import BaseIO
-from modin.experimental.engines.gandiva_on_ray.frame.partition_manager import (
-    GandivaOnRayFrameManager,
-)
-
-
-class GandivaOnRayIO(BaseIO):
-
-    block_partitions_cls = GandivaOnRayFrameManager
-    query_compiler_cls = GandivaQueryCompiler
