+++ modin/experimental/engines/pandas_on_ray/io_exp.py
@@ -69,7 +69,7 @@ class ExperimentalPandasOnRayIO(PandasOnRayIO):
             )
         #  starts the distributed alternative
         cols_names, query = get_query_info(sql, con, partition_column)
-        num_parts = cls.block_partitions_cls._compute_num_partitions()
+        num_parts = cls.frame_mgr_cls._compute_num_partitions()
         num_splits = min(len(cols_names), num_parts)
         diff = (upper_bound - lower_bound) + 1
         min_size = diff // num_parts
@@ -108,7 +108,7 @@ class ExperimentalPandasOnRayIO(PandasOnRayIO):
             index_ids.append(partition_id[-1])
         new_index = pandas.RangeIndex(sum(ray.get(index_ids)))
         new_query_compiler = cls.query_compiler_cls(
-            cls.block_partitions_cls(np.array(partition_ids)), new_index, cols_names
+            cls.frame_mgr_cls(np.array(partition_ids)), new_index, cols_names
         )
         return new_query_compiler
